+++ src/graph/blocking.rs
@@ -208,6 +208,7 @@ impl SyncGraph {
     )]
     pub fn list_indices(&mut self) -> FalkorResult<QueryResult<Vec<FalkorIndex>>> {
         ProcedureQueryBuilder::<QueryResult<Vec<FalkorIndex>>, Self>::new(self, "DB.INDEXES")
+            .read_only()
             .execute()
     }
 
@@ -276,6 +277,7 @@ impl SyncGraph {
     )]
     pub fn list_constraints(&mut self) -> FalkorResult<QueryResult<Vec<Constraint>>> {
         ProcedureQueryBuilder::<QueryResult<Vec<Constraint>>, Self>::new(self, "DB.CONSTRAINTS")
+            .read_only()
             .execute()
     }